requirements.txt
transformers
gradio
torch
torchvision
timm
diffusers
accelerate
pillow
scikit-image
matplotlib
python-dotenv
